Correct stub to return string, per the method it stubs
The comparisons made in the conditionals for the resume event are faulty,
comparing the results of `Shipment#determine_state` with symbols, when that
method returns only strings. The test only passed because the method was
stubbed to return a symbol.

Correct the stub to make the test fail properly.

Fix shipment :resume event conditional comparisons

`Spree::Shipment#determine_state returns strings, not symbols.

Correct Spree::Shipment#resume specs and remove redundant statement

Fixes #6580